pyproject.toml
[project]
name = "quiz-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-telegram-bot>=21.0,<22.0",
    "openai>=1.50.0",
    "pypdf>=4.0.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=26.3.1",
    "mypy>=1.20.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[tool.mypy]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["core", "backend", "bot"]
